Rearrange Dag parsing tests out of test_scheduler_job.py
Dag parsing code is not entirely isolated in to airflow/dag_processing/
so it makes sense to move the tests to match -- nothing in
test_scheduler_job should be dealing directly with DAG files anymore.
